feat: do alert on delete roadmap

+++ src/containers/MainDashboard/MainDashboard.tsx
@@ -96,10 +96,12 @@ const MainDashboard: React.FC = () => {
   // eslint-disable-next-line no-shadow
   const handleDelete = async (e: React.MouseEvent<HTMLButtonElement>, id: string) => {
     e.preventDefault();
-    await deleteRoadmap({
-      variables: { id },
-    });
-    refetch();
+    if (window.confirm('Are you sure you wish to delete this Roadmap?')) {
+      await deleteRoadmap({
+        variables: { id },
+      });
+      refetch();
+    }
   };
 
   // if the data is still loading
